+++ src/main/java/murskiy_sergey/graduate_work/services/analyze/AnalyzeResponse.java
@@ -5,11 +5,11 @@ import java.util.List;
 
 public class AnalyzeResponse {
     private List<AnalyzeResponseEntity> analyzeResponseEntities;
-    private int textCount;
+    private long textCount;
     private long milliseconds;
     private String comment;
 
-    public AnalyzeResponse(List<AnalyzeResponseEntity> analyzeResponseEntities, int textCount, long milliseconds) {
+    public AnalyzeResponse(List<AnalyzeResponseEntity> analyzeResponseEntities, long textCount, long milliseconds) {
         this.analyzeResponseEntities = analyzeResponseEntities;
         this.textCount = textCount;
         this.milliseconds = milliseconds;
@@ -31,11 +31,11 @@ public class AnalyzeResponse {
         this.analyzeResponseEntities = analyzeResponseEntities;
     }
 
-    public int getTextCount() {
+    public long getTextCount() {
         return textCount;
     }
 
-    public void setTextCount(int textCount) {
+    public void setTextCount(long textCount) {
         this.textCount = textCount;
     }
